家里提交

+++ app/model/kk2/province.js
@@ -5,12 +5,12 @@ module.exports = app => {
 
   const Model = app.model.define('province', {
     id: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.INTEGER(11),
       allowNull: false,
       primaryKey: true
     },
     province_id: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.INTEGER(11),
       allowNull: true
     },
     province_name: {
@@ -26,16 +26,16 @@ module.exports = app => {
       allowNull: true
     },
     status: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.INTEGER(11),
       allowNull: true,
       defaultValue: '-1'
     },
     province_id_two: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.INTEGER(11),
       allowNull: true
     },
     r_province_id: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.INTEGER(11),
       allowNull: true
     },
     r_province_name: {
